requirements.txt
python-gitlab
github3.py
urllib3
